moves supp manure requests into manure manager
