+++ RUFAS/routines/manure/manure_nutrients/manure_nutrient_manager.py
@@ -158,93 +158,6 @@ class ManureNutrientManager:
             )
 
     @staticmethod
-    def combine_manure_request_results(
-        first_request: NutrientRequestResults, second_request: NutrientRequestResults
-    ) -> NutrientRequestResults:
-        """
-        Combines the results of two manure requests.
-
-        Parameters
-        ----------
-        first_request : NutrientRequestResults
-            The results of the first nutrient request.
-        second_request : NutrientRequestResults
-            The results of the second nutrient request.
-
-        Returns
-        -------
-        NutrientRequestResults
-            The combined results of the two manure requests.
-        """
-        total_mass = first_request.total_manure_mass + second_request.total_manure_mass
-
-        def weighted_average(attr: str) -> float:
-            return (
-                (
-                    getattr(first_request, attr) * first_request.total_manure_mass
-                    + getattr(second_request, attr) * second_request.total_manure_mass
-                )
-                / total_mass
-                if total_mass > 0
-                else 0.0
-            )
-
-        return NutrientRequestResults(
-            nitrogen=first_request.nitrogen + second_request.nitrogen,
-            phosphorus=first_request.phosphorus + second_request.phosphorus,
-            total_manure_mass=total_mass,
-            organic_nitrogen_fraction=weighted_average("organic_nitrogen_fraction"),
-            inorganic_nitrogen_fraction=weighted_average("inorganic_nitrogen_fraction"),
-            ammonium_nitrogen_fraction=weighted_average("ammonium_nitrogen_fraction"),
-            organic_phosphorus_fraction=weighted_average("organic_phosphorus_fraction"),
-            inorganic_phosphorus_fraction=weighted_average("inorganic_phosphorus_fraction"),
-            dry_matter=first_request.dry_matter + second_request.dry_matter,
-            dry_matter_fraction=weighted_average("dry_matter_fraction"),
-        )
-
-    @staticmethod
-    def calculate_supplemental_manure_needed(
-        on_farm_manure: NutrientRequestResults | None, nutrient_request: NutrientRequest
-    ) -> NutrientRequest:
-        """
-        Calculate the amount of supplemental manure needed to fulfill the nutrient request.
-
-        Parameters
-        ----------
-        on_farm_manure : NutrientRequestResults | None
-            The results of the nutrient request for manure available from the farm. If None, it means that
-            there was no available on-farm manure.
-        nutrient_request : NutrientRequest
-            The nutrient request.
-
-        Returns
-        -------
-        NutrientRequest
-            The request for supplemental manure needed to fulfill the original nutrient request.
-        """
-        remaining_nitrogen = max(0, nutrient_request.nitrogen - (on_farm_manure.nitrogen if on_farm_manure else 0))
-        remaining_phosphorus = max(
-            0, nutrient_request.phosphorus - (on_farm_manure.phosphorus if on_farm_manure else 0)
-        )
-
-        if math.isclose(remaining_nitrogen, 0.0, abs_tol=1e-6) and math.isclose(
-            remaining_phosphorus, 0.0, abs_tol=1e-6
-        ):
-            return NutrientRequest(
-                nitrogen=0.0,
-                phosphorus=0.0,
-                manure_type=nutrient_request.manure_type,
-                use_supplemental_manure=True,
-            )
-
-        return NutrientRequest(
-            nitrogen=remaining_nitrogen,
-            phosphorus=remaining_phosphorus,
-            manure_type=nutrient_request.manure_type,
-            use_supplemental_manure=True,
-        )
-
-    @staticmethod
     def _calculate_projected_manure_mass(request_nutrient: float, nutrient_composition: float) -> float:
         """
         Calculate the projected manure mass based on the nutrient requested and the nutrient's composition
