+++ tests/test_manure_module/test_manure_manager.py
@@ -1,7 +1,9 @@
+import math
 import pytest
 from mock import MagicMock, call
 from pytest_mock import MockFixture, MockerFixture
 
+from RUFAS.data_structures.manure_to_crop_soil_connection import NutrientRequest, NutrientRequestResults
 from RUFAS.routines.manure.constants_and_units.manure_constants import ManureConstants
 from RUFAS.routines.manure.field_manure_supplier import FieldManureSupplier
 from RUFAS.routines.manure.IO_helpers.manure_module_output_manager_helper import ManureModuleOutputManagerHelper
@@ -844,8 +846,10 @@ def test_request_nutrients(mocker: MockerFixture, animals_simulated: bool, use_s
     # Mock dependencies
     mock_manure_nutrient_manager = mocker.MagicMock()
     mock_field_manure_supplier = mocker.MagicMock()
+    mock_output_manager = mocker.MagicMock()
     manure_manager._manure_nutrient_manager = mock_manure_nutrient_manager
     manure_manager._field_manure_supplier = mock_field_manure_supplier
+    manure_manager.om = mock_output_manager
 
     # Mock nutrient request
     mock_nutrient_request = mocker.MagicMock()
@@ -856,10 +860,8 @@ def test_request_nutrients(mocker: MockerFixture, animals_simulated: bool, use_s
     # Patch methods
     mocker.patch.object(mock_manure_nutrient_manager, "request_nutrients", return_value=mock_request_nutrients_result)
     mocker.patch.object(mock_field_manure_supplier, "request_nutrients", return_value=mocker.MagicMock())
-    mocker.patch.object(
-        mock_manure_nutrient_manager, "calculate_supplemental_manure_needed", return_value=mocker.MagicMock()
-    )
-    mocker.patch.object(mock_manure_nutrient_manager, "combine_manure_request_results", return_value=mocker.MagicMock())
+    mocker.patch.object(manure_manager, "_calculate_supplemental_manure_needed", return_value=mocker.MagicMock())
+    mocker.patch.object(manure_manager, "_combine_manure_request_results", return_value=mocker.MagicMock())
     mocker.patch.object(manure_manager, "_record_manure_request_results")
 
     # Act
@@ -872,17 +874,22 @@ def test_request_nutrients(mocker: MockerFixture, animals_simulated: bool, use_s
         if not use_supplemental_manure:
             assert actual_results == mock_nutrient_request_results
         else:
-            mock_manure_nutrient_manager.calculate_supplemental_manure_needed.assert_called_once_with(
+            mock_output_manager.add_log.assert_called_once_with(
+                "Supplemental manure needed",
+                "Attempting to fulfill manure nutrient request shortfall with supplemental manure.",
+                {"class": manure_manager.__class__.__name__, "function": manure_manager.request_nutrients.__name__},
+            )
+            manure_manager._calculate_supplemental_manure_needed.assert_called_once_with(
                 mock_nutrient_request_results, mock_nutrient_request
             )
             mock_field_manure_supplier.request_nutrients.assert_called_once()
             manure_manager._record_manure_request_results.assert_any_call(
                 mock_field_manure_supplier.request_nutrients(), "supplemental_manure"
             )
-            mock_manure_nutrient_manager.combine_manure_request_results.assert_called_once_with(
+            manure_manager._combine_manure_request_results.assert_called_once_with(
                 mock_nutrient_request_results, mock_field_manure_supplier.request_nutrients()
             )
-            assert actual_results == mock_manure_nutrient_manager.combine_manure_request_results()
+            assert actual_results == manure_manager._combine_manure_request_results()
     else:
         mock_field_manure_supplier.request_nutrients.assert_called_once_with(mock_nutrient_request)
         assert actual_results == mock_field_manure_supplier.request_nutrients()
@@ -906,18 +913,25 @@ def test_record_manure_request_results(mocker: MockerFixture) -> None:
         nitrogen=50.0,
         phosphorus=10.0,
     )
+    mocker.patch("RUFAS.routines.manure.manure_manager.ManureManager.__init__", return_value=None)
+    manure_manager = ManureManager(
+        animal_manager=mocker.MagicMock(),
+        weather=mocker.MagicMock(),
+        time=mocker.MagicMock(),
+        manure_manager_config=mocker.MagicMock(),
+        simulate_animals=True,
+    )
 
-    mock_output_manager = mocker.patch("RUFAS.routines.manure.manure_manager.OutputManager")
-    mock_add_variable = mock_output_manager.return_value.add_variable
+    mock_output_manager = mocker.MagicMock()
+    manure_manager.om = mock_output_manager
 
     # Act
-    ManureManager._record_manure_request_results(mock_nutrient_request_results, manure_source)
+    manure_manager._record_manure_request_results(mock_nutrient_request_results, manure_source)
 
     # Assert
-    mock_add_variable.assert_called_once()
-    actual_manure_source, actual_request_result_values, actual_info_maps = mock_add_variable.call_args[0]
+    mock_output_manager.add_variable.assert_called_once()
+    actual_manure_source, actual_request_result_values, actual_info_maps = mock_output_manager.add_variable.call_args[0]
     assert actual_manure_source == manure_source
-
     expected_request_result_values = {
         "dry_matter_mass": mock_nutrient_request_results.dry_matter,
         "dry_matter_fraction": mock_nutrient_request_results.dry_matter_fraction,
@@ -949,3 +963,231 @@ def test_record_manure_request_results(mocker: MockerFixture) -> None:
         },
     }
     assert actual_info_maps == expected_info_maps
+
+
+@pytest.mark.parametrize(
+    "first_request, second_request, expected_result",
+    [
+        # Scenario: Combining two requests with valid fractions
+        (
+            NutrientRequestResults(
+                nitrogen=8,
+                phosphorus=6,
+                total_manure_mass=10,
+                organic_nitrogen_fraction=0.4,
+                inorganic_nitrogen_fraction=0.6,
+                ammonium_nitrogen_fraction=0.3,
+                organic_phosphorus_fraction=0.5,
+                inorganic_phosphorus_fraction=0.5,
+                dry_matter=2,
+                dry_matter_fraction=0.1,
+            ),
+            NutrientRequestResults(
+                nitrogen=12,
+                phosphorus=9,
+                total_manure_mass=20,
+                organic_nitrogen_fraction=0.5,
+                inorganic_nitrogen_fraction=0.5,
+                ammonium_nitrogen_fraction=0.4,
+                organic_phosphorus_fraction=0.4,
+                inorganic_phosphorus_fraction=0.6,
+                dry_matter=3,
+                dry_matter_fraction=0.15,
+            ),
+            NutrientRequestResults(
+                nitrogen=20,
+                phosphorus=15,
+                total_manure_mass=30,
+                organic_nitrogen_fraction=0.4667,
+                inorganic_nitrogen_fraction=0.5333,
+                ammonium_nitrogen_fraction=0.3667,
+                organic_phosphorus_fraction=0.4333,
+                inorganic_phosphorus_fraction=0.5667,
+                dry_matter=5,
+                dry_matter_fraction=0.1333,
+            ),
+        ),
+        # Scenario: One request has zero total manure mass
+        (
+            NutrientRequestResults(
+                nitrogen=0,
+                phosphorus=0,
+                total_manure_mass=0,
+                organic_nitrogen_fraction=0.7,
+                inorganic_nitrogen_fraction=0.3,
+                ammonium_nitrogen_fraction=0.5,
+                organic_phosphorus_fraction=0.6,
+                inorganic_phosphorus_fraction=0.4,
+                dry_matter=0,
+                dry_matter_fraction=0.0,
+            ),
+            NutrientRequestResults(
+                nitrogen=10,
+                phosphorus=5,
+                total_manure_mass=15,
+                organic_nitrogen_fraction=0.6,
+                inorganic_nitrogen_fraction=0.4,
+                ammonium_nitrogen_fraction=0.3,
+                organic_phosphorus_fraction=0.5,
+                inorganic_phosphorus_fraction=0.5,
+                dry_matter=3,
+                dry_matter_fraction=0.2,
+            ),
+            NutrientRequestResults(
+                nitrogen=10,
+                phosphorus=5,
+                total_manure_mass=15,
+                organic_nitrogen_fraction=0.6,
+                inorganic_nitrogen_fraction=0.4,
+                ammonium_nitrogen_fraction=0.3,
+                organic_phosphorus_fraction=0.5,
+                inorganic_phosphorus_fraction=0.5,
+                dry_matter=3,
+                dry_matter_fraction=0.2,
+            ),
+        ),
+    ],
+)
+def test_combine_manure_request_results(
+    first_request: NutrientRequestResults,
+    second_request: NutrientRequestResults,
+    expected_result: NutrientRequestResults,
+    mocker: MockFixture,
+) -> None:
+    """
+    Unit test for the _combine_manure_request_results static method.
+    """
+    # Arrange
+    mocker.patch("RUFAS.routines.manure.manure_manager.ManureManager.__init__", return_value=None)
+    manure_manager = ManureManager(
+        animal_manager=mocker.MagicMock(),
+        weather=mocker.MagicMock(),
+        time=mocker.MagicMock(),
+        manure_manager_config=mocker.MagicMock(),
+        simulate_animals=True,
+    )
+    # Act
+    actual_result = manure_manager._combine_manure_request_results(first_request, second_request)
+
+    # Assert
+    assert actual_result.nitrogen == expected_result.nitrogen
+    assert actual_result.phosphorus == expected_result.phosphorus
+    assert actual_result.total_manure_mass == expected_result.total_manure_mass
+    assert math.isclose(
+        actual_result.organic_nitrogen_fraction, expected_result.organic_nitrogen_fraction, abs_tol=1e-4
+    )
+    assert math.isclose(
+        actual_result.inorganic_nitrogen_fraction, expected_result.inorganic_nitrogen_fraction, abs_tol=1e-4
+    )
+    assert math.isclose(
+        actual_result.ammonium_nitrogen_fraction, expected_result.ammonium_nitrogen_fraction, abs_tol=1e-4
+    )
+    assert math.isclose(
+        actual_result.organic_phosphorus_fraction, expected_result.organic_phosphorus_fraction, abs_tol=1e-4
+    )
+    assert math.isclose(
+        actual_result.inorganic_phosphorus_fraction, expected_result.inorganic_phosphorus_fraction, abs_tol=1e-4
+    )
+    assert actual_result.dry_matter == expected_result.dry_matter
+    assert math.isclose(actual_result.dry_matter_fraction, expected_result.dry_matter_fraction, abs_tol=1e-4)
+
+
+@pytest.mark.parametrize(
+    "on_farm_manure, nutrient_request, expected_result",
+    [
+        # Scenario: No supplemental manure needed (on-farm manure fully satisfies the request)
+        (
+            NutrientRequestResults(
+                nitrogen=10,
+                phosphorus=5,
+                total_manure_mass=15,
+                organic_nitrogen_fraction=0.6,
+                inorganic_nitrogen_fraction=0.4,
+                ammonium_nitrogen_fraction=0.3,
+                organic_phosphorus_fraction=0.5,
+                inorganic_phosphorus_fraction=0.5,
+                dry_matter=3,
+                dry_matter_fraction=0.2,
+            ),
+            NutrientRequest(
+                nitrogen=8,
+                phosphorus=4,
+                manure_type=ManureType.LIQUID,
+                use_supplemental_manure=False,
+            ),
+            NutrientRequest(
+                nitrogen=0.0,
+                phosphorus=0.0,
+                manure_type=ManureType.LIQUID,
+                use_supplemental_manure=True,
+            ),
+        ),
+        # Scenario: Partial supplemental manure needed (on-farm manure partially satisfies the request)
+        (
+            NutrientRequestResults(
+                nitrogen=5,
+                phosphorus=2,
+                total_manure_mass=10,
+                organic_nitrogen_fraction=0.7,
+                inorganic_nitrogen_fraction=0.3,
+                ammonium_nitrogen_fraction=0.5,
+                organic_phosphorus_fraction=0.6,
+                inorganic_phosphorus_fraction=0.4,
+                dry_matter=2,
+                dry_matter_fraction=0.1,
+            ),
+            NutrientRequest(
+                nitrogen=8,
+                phosphorus=5,
+                manure_type=ManureType.LIQUID,
+                use_supplemental_manure=False,
+            ),
+            NutrientRequest(
+                nitrogen=3.0,
+                phosphorus=3.0,
+                manure_type=ManureType.LIQUID,
+                use_supplemental_manure=True,
+            ),
+        ),
+        # Scenario: All supplemental manure needed (on-farm manure provides nothing)
+        (
+            None,
+            NutrientRequest(
+                nitrogen=10,
+                phosphorus=6,
+                manure_type=ManureType.LIQUID,
+                use_supplemental_manure=False,
+            ),
+            NutrientRequest(
+                nitrogen=10.0,
+                phosphorus=6.0,
+                manure_type=ManureType.LIQUID,
+                use_supplemental_manure=True,
+            ),
+        ),
+    ],
+)
+def test_calculate_supplemental_manure_needed(
+    on_farm_manure: NutrientRequestResults, nutrient_request: NutrientRequest, expected_result: NutrientRequest,
+    mocker: MockFixture,
+) -> None:
+    """
+    Unit test for the _calculate_supplemental_manure_needed static method.
+    """
+    # Arrange
+    mocker.patch("RUFAS.routines.manure.manure_manager.ManureManager.__init__", return_value=None)
+    manure_manager = ManureManager(
+        animal_manager=mocker.MagicMock(),
+        weather=mocker.MagicMock(),
+        time=mocker.MagicMock(),
+        manure_manager_config=mocker.MagicMock(),
+        simulate_animals=True,
+    )
+    # Act
+    actual_result = manure_manager._calculate_supplemental_manure_needed(on_farm_manure, nutrient_request)
+
+    # Assert
+    assert math.isclose(actual_result.nitrogen, expected_result.nitrogen, abs_tol=1e-6)
+    assert math.isclose(actual_result.phosphorus, expected_result.phosphorus, abs_tol=1e-6)
+    assert actual_result.manure_type == expected_result.manure_type
+    assert actual_result.use_supplemental_manure == expected_result.use_supplemental_manure
