+++ tests/test_manure_module/test_manure_nutrient_manager.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import math
 import re
 
 import pytest
@@ -292,214 +291,6 @@ def test_evaluate_nutrient_request(
 
     assert actual_result == expected_request_result
     assert actual_fulfilled == expected_fulfilled
-
-
-@pytest.mark.parametrize(
-    "first_request, second_request, expected_result",
-    [
-        # Scenario: Combining two requests with valid fractions
-        (
-            NutrientRequestResults(
-                nitrogen=8,
-                phosphorus=6,
-                total_manure_mass=10,
-                organic_nitrogen_fraction=0.4,
-                inorganic_nitrogen_fraction=0.6,
-                ammonium_nitrogen_fraction=0.3,
-                organic_phosphorus_fraction=0.5,
-                inorganic_phosphorus_fraction=0.5,
-                dry_matter=2,
-                dry_matter_fraction=0.1,
-            ),
-            NutrientRequestResults(
-                nitrogen=12,
-                phosphorus=9,
-                total_manure_mass=20,
-                organic_nitrogen_fraction=0.5,
-                inorganic_nitrogen_fraction=0.5,
-                ammonium_nitrogen_fraction=0.4,
-                organic_phosphorus_fraction=0.4,
-                inorganic_phosphorus_fraction=0.6,
-                dry_matter=3,
-                dry_matter_fraction=0.15,
-            ),
-            NutrientRequestResults(
-                nitrogen=20,
-                phosphorus=15,
-                total_manure_mass=30,
-                organic_nitrogen_fraction=0.4667,
-                inorganic_nitrogen_fraction=0.5333,
-                ammonium_nitrogen_fraction=0.3667,
-                organic_phosphorus_fraction=0.4333,
-                inorganic_phosphorus_fraction=0.5667,
-                dry_matter=5,
-                dry_matter_fraction=0.1333,
-            ),
-        ),
-        # Scenario: One request has zero total manure mass
-        (
-            NutrientRequestResults(
-                nitrogen=0,
-                phosphorus=0,
-                total_manure_mass=0,
-                organic_nitrogen_fraction=0.7,
-                inorganic_nitrogen_fraction=0.3,
-                ammonium_nitrogen_fraction=0.5,
-                organic_phosphorus_fraction=0.6,
-                inorganic_phosphorus_fraction=0.4,
-                dry_matter=0,
-                dry_matter_fraction=0.0,
-            ),
-            NutrientRequestResults(
-                nitrogen=10,
-                phosphorus=5,
-                total_manure_mass=15,
-                organic_nitrogen_fraction=0.6,
-                inorganic_nitrogen_fraction=0.4,
-                ammonium_nitrogen_fraction=0.3,
-                organic_phosphorus_fraction=0.5,
-                inorganic_phosphorus_fraction=0.5,
-                dry_matter=3,
-                dry_matter_fraction=0.2,
-            ),
-            NutrientRequestResults(
-                nitrogen=10,
-                phosphorus=5,
-                total_manure_mass=15,
-                organic_nitrogen_fraction=0.6,
-                inorganic_nitrogen_fraction=0.4,
-                ammonium_nitrogen_fraction=0.3,
-                organic_phosphorus_fraction=0.5,
-                inorganic_phosphorus_fraction=0.5,
-                dry_matter=3,
-                dry_matter_fraction=0.2,
-            ),
-        ),
-    ],
-)
-def test_combine_manure_request_results(
-    first_request: NutrientRequestResults,
-    second_request: NutrientRequestResults,
-    expected_result: NutrientRequestResults,
-) -> None:
-    """
-    Unit test for the _combine_manure_request_results static method.
-    """
-    # Act
-    actual_result = ManureNutrientManager.combine_manure_request_results(first_request, second_request)
-
-    # Assert
-    assert actual_result.nitrogen == expected_result.nitrogen
-    assert actual_result.phosphorus == expected_result.phosphorus
-    assert actual_result.total_manure_mass == expected_result.total_manure_mass
-    assert math.isclose(
-        actual_result.organic_nitrogen_fraction, expected_result.organic_nitrogen_fraction, abs_tol=1e-4
-    )
-    assert math.isclose(
-        actual_result.inorganic_nitrogen_fraction, expected_result.inorganic_nitrogen_fraction, abs_tol=1e-4
-    )
-    assert math.isclose(
-        actual_result.ammonium_nitrogen_fraction, expected_result.ammonium_nitrogen_fraction, abs_tol=1e-4
-    )
-    assert math.isclose(
-        actual_result.organic_phosphorus_fraction, expected_result.organic_phosphorus_fraction, abs_tol=1e-4
-    )
-    assert math.isclose(
-        actual_result.inorganic_phosphorus_fraction, expected_result.inorganic_phosphorus_fraction, abs_tol=1e-4
-    )
-    assert actual_result.dry_matter == expected_result.dry_matter
-    assert math.isclose(actual_result.dry_matter_fraction, expected_result.dry_matter_fraction, abs_tol=1e-4)
-
-
-@pytest.mark.parametrize(
-    "on_farm_manure, nutrient_request, expected_result",
-    [
-        # Scenario: No supplemental manure needed (on-farm manure fully satisfies the request)
-        (
-            NutrientRequestResults(
-                nitrogen=10,
-                phosphorus=5,
-                total_manure_mass=15,
-                organic_nitrogen_fraction=0.6,
-                inorganic_nitrogen_fraction=0.4,
-                ammonium_nitrogen_fraction=0.3,
-                organic_phosphorus_fraction=0.5,
-                inorganic_phosphorus_fraction=0.5,
-                dry_matter=3,
-                dry_matter_fraction=0.2,
-            ),
-            NutrientRequest(
-                nitrogen=8,
-                phosphorus=4,
-                manure_type=ManureType.LIQUID,
-                use_supplemental_manure=False,
-            ),
-            NutrientRequest(
-                nitrogen=0.0,
-                phosphorus=0.0,
-                manure_type=ManureType.LIQUID,
-                use_supplemental_manure=True,
-            ),
-        ),
-        # Scenario: Partial supplemental manure needed (on-farm manure partially satisfies the request)
-        (
-            NutrientRequestResults(
-                nitrogen=5,
-                phosphorus=2,
-                total_manure_mass=10,
-                organic_nitrogen_fraction=0.7,
-                inorganic_nitrogen_fraction=0.3,
-                ammonium_nitrogen_fraction=0.5,
-                organic_phosphorus_fraction=0.6,
-                inorganic_phosphorus_fraction=0.4,
-                dry_matter=2,
-                dry_matter_fraction=0.1,
-            ),
-            NutrientRequest(
-                nitrogen=8,
-                phosphorus=5,
-                manure_type=ManureType.LIQUID,
-                use_supplemental_manure=False,
-            ),
-            NutrientRequest(
-                nitrogen=3.0,
-                phosphorus=3.0,
-                manure_type=ManureType.LIQUID,
-                use_supplemental_manure=True,
-            ),
-        ),
-        # Scenario: All supplemental manure needed (on-farm manure provides nothing)
-        (
-            None,
-            NutrientRequest(
-                nitrogen=10,
-                phosphorus=6,
-                manure_type=ManureType.LIQUID,
-                use_supplemental_manure=False,
-            ),
-            NutrientRequest(
-                nitrogen=10.0,
-                phosphorus=6.0,
-                manure_type=ManureType.LIQUID,
-                use_supplemental_manure=True,
-            ),
-        ),
-    ],
-)
-def test_calculate_supplemental_manure_needed(
-    on_farm_manure: NutrientRequestResults, nutrient_request: NutrientRequest, expected_result: NutrientRequest
-) -> None:
-    """
-    Unit test for the _calculate_supplemental_manure_needed static method.
-    """
-    # Act
-    actual_result = ManureNutrientManager.calculate_supplemental_manure_needed(on_farm_manure, nutrient_request)
-
-    # Assert
-    assert math.isclose(actual_result.nitrogen, expected_result.nitrogen, abs_tol=1e-6)
-    assert math.isclose(actual_result.phosphorus, expected_result.phosphorus, abs_tol=1e-6)
-    assert actual_result.manure_type == expected_result.manure_type
-    assert actual_result.use_supplemental_manure == expected_result.use_supplemental_manure
 
 
 @pytest.mark.parametrize(
